feat: enahance UX for both create and start
