+++ packages/cli/src/commands/create.ts
@@ -13,6 +13,7 @@ import prompts from 'prompts';
 import colors from 'yoctocolors';
 import { z } from 'zod';
 import { displayBanner } from '../displayBanner';
+import { setupPgLite, promptAndStorePostgresUrl, getElizaDirectories } from '../utils/get-config';
 
 /**
  * This module handles creating both projects and plugins.
@@ -78,154 +79,6 @@ async function installDependencies(targetDir: string) {
   } catch (error) {
     logger.warn("Failed to install dependencies automatically. Please run 'bun install' manually.");
   }
-}
-
-/**
- * Stores Postgres URL in the global .env file
- * @param url The Postgres URL to store
- */
-async function storePostgresUrl(url: string): Promise<void> {
-  if (!url) return;
-
-  try {
-    const homeDir = os.homedir();
-    const globalEnvPath = path.join(homeDir, '.eliza', '.env');
-
-    // Ensure .eliza directory exists
-    const elizaDir = path.join(homeDir, '.eliza');
-    if (!existsSync(elizaDir)) {
-      await fs.mkdir(elizaDir, { recursive: true });
-    }
-
-    // Create .env file if it doesn't exist
-    if (!existsSync(globalEnvPath)) {
-      await fs.writeFile(globalEnvPath, '', { encoding: 'utf8' });
-    }
-
-    await fs.writeFile(globalEnvPath, `POSTGRES_URL=${url}\n`, { flag: 'a' });
-
-    // Also set in process.env for the current session
-    process.env.POSTGRES_URL = url;
-
-    logger.success('Postgres URL saved to configuration');
-  } catch (error) {
-    logger.warn('Error saving database configuration:', error);
-  }
-}
-
-/**
- * Sets up and configures PGLite database
- * @param elizaDbDir The directory for PGLite database
- */
-async function setupPgLiteDir(elizaDbDir: string): Promise<void> {
-  try {
-    const homeDir = os.homedir();
-    const elizaDir = path.join(homeDir, '.eliza');
-    const envFilePath = path.join(elizaDir, '.env');
-
-    // Ensure .eliza directory exists
-    if (!existsSync(elizaDir)) {
-      await fs.mkdir(elizaDir, { recursive: true });
-      logger.info(`Created directory: ${elizaDir}`);
-    }
-
-    // Ensure the PGLite database directory exists
-    if (!existsSync(elizaDbDir)) {
-      await fs.mkdir(elizaDbDir, { recursive: true });
-      logger.info(`Created PGLite database directory: ${elizaDbDir}`);
-    }
-
-    // Create or update .env file
-    if (!existsSync(envFilePath)) {
-      await fs.writeFile(envFilePath, '', { encoding: 'utf8' });
-    }
-
-    // Store PGLITE_DATA_DIR in the environment file
-    await fs.writeFile(envFilePath, `PGLITE_DATA_DIR=${elizaDbDir}\n`, { flag: 'a' });
-
-    // Also set in process.env for the current session
-    process.env.PGLITE_DATA_DIR = elizaDbDir;
-
-    logger.success('PGLite configuration saved');
-  } catch (error) {
-    logger.error('Error setting up PGLite directory:', error);
-    throw error;
-  }
-}
-
-/**
- * Validates a Postgres URL format
- * @param url The URL to validate
- * @returns True if the URL appears valid
- */
-function isValidPostgresUrl(url: string): boolean {
-  if (!url) return false;
-
-  // Basic pattern: postgresql://user:password@host:port/dbname
-  const basicPattern = /^postgresql:\/\/[^:]+:[^@]+@[^:]+:\d+\/\w+$/;
-
-  // More permissive pattern (allows missing password, different formats)
-  const permissivePattern = /^postgresql:\/\/.*@.*:\d+\/.*$/;
-
-  return basicPattern.test(url) || permissivePattern.test(url);
-}
-
-/**
- * Prompts the user for a Postgres URL, validates it, and stores it
- * @returns The configured Postgres URL or null if user skips
- */
-async function promptAndStorePostgresUrl(): Promise<string | null> {
-  let isValidUrl = false;
-  let userUrl = '';
-
-  while (!isValidUrl) {
-    // Prompt for postgres url with simpler message
-    const reply = await prompts({
-      type: 'text',
-      name: 'postgresUrl',
-      message: 'Enter your Postgres URL:',
-      validate: (value) => value.trim() !== '' || 'Postgres URL cannot be empty',
-    });
-
-    // Handle cancellation
-    if (!reply.postgresUrl) {
-      const { continueAnyway } = await prompts({
-        type: 'confirm',
-        name: 'continueAnyway',
-        message: 'Continue without configuring Postgres?',
-        initial: false,
-      });
-
-      if (continueAnyway) return null;
-      continue;
-    }
-
-    userUrl = reply.postgresUrl;
-
-    // Validate URL format
-    if (!isValidPostgresUrl(userUrl)) {
-      logger.warn("The URL format doesn't appear to be valid.");
-      logger.info('Expected format: postgresql://user:password@host:port/dbname');
-
-      const { useAnyway } = await prompts({
-        type: 'confirm',
-        name: 'useAnyway',
-        message: 'Use this URL anyway? (Choose Yes if you have a custom setup)',
-        initial: false,
-      });
-
-      if (!useAnyway) continue;
-    }
-
-    isValidUrl = true;
-  }
-
-  if (userUrl) {
-    await storePostgresUrl(userUrl);
-    return userUrl;
-  }
-
-  return null;
 }
 
 /**
@@ -407,18 +260,16 @@ export const create = new Command()
       await copyTemplate('project', targetDir, name);
 
       // Database configuration
-      const homeDir = os.homedir();
-      const elizaDir = path.join(homeDir, '.eliza');
-      const elizaDbDir = path.join(elizaDir, 'db');
+      const { elizaDir, elizaDbDir, envFilePath } = getElizaDirectories();
 
       // Only create directories and configure based on database choice
       if (database === 'pglite') {
         // Set up PGLite directory and configuration
-        await setupPgLiteDir(elizaDbDir);
+        await setupPgLite(elizaDbDir, envFilePath);
         logger.debug(`Using PGLite database directory: ${elizaDbDir}`);
       } else if (database === 'postgres' && !postgresUrl) {
         // Handle Postgres configuration
-        postgresUrl = await promptAndStorePostgresUrl();
+        postgresUrl = await promptAndStorePostgresUrl(envFilePath);
       }
 
       // Set up src directory
