+++ packages/cli/src/utils/get-config.ts
@@ -3,6 +3,8 @@ import os from 'node:os';
 import path from 'node:path';
 import dotenv from 'dotenv';
 import { z } from 'zod';
+import prompts from 'prompts';
+import { logger } from '@elizaos/core';
 
 // Database config schemas
 const postgresConfigSchema = z.object({
@@ -18,6 +20,238 @@ const pgliteConfigSchema = z.object({
     dataDir: z.string(),
   }),
 });
+
+/**
+ * Validates a Postgres URL format
+ * @param url The URL to validate
+ * @returns True if the URL appears valid
+ */
+export function isValidPostgresUrl(url: string): boolean {
+  if (!url) return false;
+
+  // Basic pattern: postgresql://user:password@host:port/dbname
+  const basicPattern = /^postgresql:\/\/[^:]+:[^@]+@[^:]+:\d+\/\w+$/;
+
+  // More permissive pattern (allows missing password, different formats)
+  const permissivePattern = /^postgresql:\/\/.*@.*:\d+\/.*$/;
+
+  return basicPattern.test(url) || permissivePattern.test(url);
+}
+
+/**
+ * Gets the standard Eliza directories
+ * @returns Object containing standard directory paths
+ */
+export function getElizaDirectories() {
+  const homeDir = os.homedir();
+  const elizaDir = path.join(homeDir, '.eliza');
+  const elizaDbDir = path.join(elizaDir, 'db');
+  const envFilePath = path.join(elizaDir, '.env');
+
+  return {
+    homeDir,
+    elizaDir,
+    elizaDbDir,
+    envFilePath,
+  };
+}
+
+/**
+ * Ensures the .eliza directory exists
+ * @returns The eliza directories object
+ */
+export async function ensureElizaDir() {
+  const dirs = getElizaDirectories();
+
+  if (!existsSync(dirs.elizaDir)) {
+    await fs.mkdir(dirs.elizaDir, { recursive: true });
+    logger.info(`Created directory: ${dirs.elizaDir}`);
+  }
+
+  return dirs;
+}
+
+/**
+ * Ensures the .env file exists
+ * @param envFilePath Path to the .env file
+ */
+export async function ensureEnvFile(envFilePath: string) {
+  if (!existsSync(envFilePath)) {
+    await fs.writeFile(envFilePath, '', { encoding: 'utf8' });
+    logger.debug(`Created empty .env file at ${envFilePath}`);
+  }
+}
+
+/**
+ * Sets up and configures PGLite database
+ * @param elizaDbDir The directory for PGLite database
+ * @param envFilePath Path to the .env file
+ */
+export async function setupPgLite(elizaDbDir: string, envFilePath: string): Promise<void> {
+  try {
+    // Ensure the PGLite database directory exists
+    if (!existsSync(elizaDbDir)) {
+      await fs.mkdir(elizaDbDir, { recursive: true });
+      logger.info(`Created PGLite database directory: ${elizaDbDir}`);
+    }
+
+    // Ensure .env file exists
+    await ensureEnvFile(envFilePath);
+
+    // Store PGLITE_DATA_DIR in the environment file
+    await fs.writeFile(envFilePath, `PGLITE_DATA_DIR=${elizaDbDir}\n`, { flag: 'a' });
+
+    // Also set in process.env for the current session
+    process.env.PGLITE_DATA_DIR = elizaDbDir;
+
+    logger.success('PGLite configuration saved');
+  } catch (error) {
+    logger.error('Error setting up PGLite directory:', error);
+    throw error;
+  }
+}
+
+/**
+ * Stores Postgres URL in the .env file
+ * @param url The Postgres URL to store
+ * @param envFilePath Path to the .env file
+ */
+export async function storePostgresUrl(url: string, envFilePath: string): Promise<void> {
+  if (!url) return;
+
+  try {
+    // Ensure .env file exists
+    await ensureEnvFile(envFilePath);
+
+    // Store the URL in the .env file
+    await fs.writeFile(envFilePath, `POSTGRES_URL=${url}\n`, { flag: 'a' });
+
+    // Also set in process.env for the current session
+    process.env.POSTGRES_URL = url;
+
+    logger.success('Postgres URL saved to configuration');
+  } catch (error) {
+    logger.warn('Error saving database configuration:', error);
+  }
+}
+
+/**
+ * Prompts the user for a Postgres URL, validates it, and stores it
+ * @param envFilePath Path to the .env file
+ * @returns The configured Postgres URL or null if user skips
+ */
+export async function promptAndStorePostgresUrl(envFilePath: string): Promise<string | null> {
+  const { postgresUrlInput } = await prompts({
+    type: 'text',
+    name: 'postgresUrlInput',
+    message: 'Enter your Postgres URL:',
+    validate: (value) => value.trim() !== '' || 'Postgres URL cannot be empty',
+  });
+
+  if (!postgresUrlInput || postgresUrlInput.trim() === '') {
+    return null;
+  }
+
+  // Basic validation
+  if (!isValidPostgresUrl(postgresUrlInput)) {
+    logger.warn("The URL format doesn't appear to be valid.");
+    logger.info('Expected format: postgresql://user:password@host:port/dbname');
+
+    const { useAnyway } = await prompts({
+      type: 'confirm',
+      name: 'useAnyway',
+      message: 'Use this URL anyway? (Choose Yes if you have a custom setup)',
+      initial: false,
+    });
+
+    if (!useAnyway) {
+      return null;
+    }
+  }
+
+  // Store the URL in the .env file
+  await storePostgresUrl(postgresUrlInput, envFilePath);
+  return postgresUrlInput;
+}
+
+/**
+ * Configures the database to use, either PGLite or PostgreSQL
+ * @returns The postgres URL if using Postgres, otherwise null
+ */
+export async function configureDatabaseSettings(): Promise<string | null> {
+  // Set up directories and env file
+  const { elizaDir, elizaDbDir, envFilePath } = await ensureElizaDir();
+  await ensureEnvFile(envFilePath);
+
+  // Check if we already have database configuration in env
+  let postgresUrl = process.env.POSTGRES_URL;
+  const pgliteDataDir = process.env.PGLITE_DATA_DIR;
+
+  // If we already have a postgres URL configured, use that
+  if (postgresUrl) {
+    logger.debug('Using existing PostgreSQL configuration');
+    return postgresUrl;
+  }
+
+  // If we already have PGLITE_DATA_DIR set in env, use PGLite
+  if (pgliteDataDir) {
+    logger.debug(`Using existing PGLite configuration: ${pgliteDataDir}`);
+
+    // Ensure the directory exists
+    if (!existsSync(pgliteDataDir)) {
+      await fs.mkdir(pgliteDataDir, { recursive: true });
+      logger.info(`Created PGLite database directory: ${pgliteDataDir}`);
+    }
+
+    return null;
+  }
+
+  try {
+    // Prompt for database selection
+    const { database } = await prompts({
+      type: 'select',
+      name: 'database',
+      message: 'Select your database:',
+      choices: [
+        { title: 'pglite (embedded database)', value: 'pglite' },
+        { title: 'postgres (external database)', value: 'postgres' },
+      ],
+      initial: 0,
+    });
+
+    if (!database || database === 'pglite') {
+      // If selection canceled or pglite selected
+      const dbChoice = !database ? 'Selection canceled, defaulting to' : 'Selected';
+      logger.info(`${dbChoice} pglite database`);
+
+      await setupPgLite(elizaDbDir, envFilePath);
+      return null;
+    }
+
+    // User selected postgres
+    const result = await promptAndStorePostgresUrl(envFilePath);
+    if (!result) {
+      // If no valid Postgres URL provided, default to PGLite
+      logger.warn('No valid Postgres URL provided, defaulting to pglite database');
+      await setupPgLite(elizaDbDir, envFilePath);
+      return null;
+    }
+
+    return result;
+  } catch (error) {
+    logger.error('Error during database configuration:', error);
+    logger.info('Defaulting to pglite database');
+
+    try {
+      await setupPgLite(elizaDbDir, envFilePath);
+    } catch (setupError) {
+      logger.error('Critical error setting up database:', setupError);
+      throw new Error('Failed to configure database');
+    }
+  }
+
+  return null; // Default to pglite
+}
 
 // Main config schema
 /**
